+++ BeaconManagement/src/main/java/beaconManagement/tcc/service/BeaconEventService.java
@@ -1,5 +1,6 @@
 package beaconManagement.tcc.service;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 import beaconManagement.tcc.domain.BeaconEvent;
@@ -14,7 +15,7 @@ public interface BeaconEventService {
 
 	List<BeaconEvent> getBeaconEvents();
 
-	List<BeaconEvent> getBeaconEventsByStartDate(Long dateMillis);
+	List<BeaconEvent> getBeaconEventsByStartDate(BigDecimal dateMillis);
 
-	List<BeaconEvent> getBeaconEventsByEndDate(Long dateMillis);
+	List<BeaconEvent> getBeaconEventsByEndDate(BigDecimal dateMillis);
 }
